# apps/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample swagger for Fiber",
        "title": "Fiber Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "fiber@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3002",
    "basePath": "/",
    "paths": {
...
        "/v1/channels/{channelId}/settings/dashboard-access": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard Access"
                ],
                "summary": "Get channel dashboard access list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ChannelId",
                        "name": "channelId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_satont_tsuwari_apps_api_internal_api_v1_settings_dashboardAccess.Entity"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_satont_tsuwari_apps_api_internal_types.DOCApiInternalError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard Access"
                ],
                "summary": "Add user to dashboard access list",
                "parameters": [
                    {
                        "description": "Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_satont_tsuwari_apps_api_internal_api_v1_settings_dashboardAccess.addUserDto"
                        }
                    },
                    {
                        "type": "string",
                        "description": "ID of channel",
                        "name": "channelId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_satont_tsuwari_apps_api_internal_api_v1_settings_dashboardAccess.Entity"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_satont_tsuwari_apps_api_internal_types.DOCApiValidationError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_satont_tsuwari_apps_api_internal_types.DOCApiInternalError"
                        }
                    }
                }
            }
        },
...
        "/v1/channels/{channelId}/timers": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Timers"
                ],
                "summary": "Get channel timers list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ChannelId",
                        "name": "channelId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.ChannelsTimers"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_satont_tsuwari_apps_api_internal_types.DOCApiInternalError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Timers"
                ],
                "summary": "Create timer",
                "parameters": [
                    {
                        "description": "Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_v1_timers.timerDto"
                        }
                    },
                    {
                        "type": "string",
                        "description": "ID of channel",
                        "name": "channelId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ChannelsTimers"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_satont_tsuwari_apps_api_internal_types.DOCApiValidationError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_satont_tsuwari_apps_api_internal_types.DOCApiInternalError"
                        }
                    }
                }
            }
        },
        "/v1/channels/{channelId}/timers/{timerId}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Timers"
                ],
                "summary": "Update timer",
                "parameters": [
                    {
                        "description": "Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_v1_timers.timerDto"
                        }
                    },
                    {
                        "type": "string",
                        "description": "ID of channel",
                        "name": "channelId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of timer",
                        "name": "timerId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ChannelsTimers"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_satont_tsuwari_apps_api_internal_types.DOCApiValidationError"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_satont_tsuwari_apps_api_internal_types.DOCApiInternalError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Timers"
                ],
                "summary": "Delete timer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of channel",
                        "name": "channelId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of timer",
                        "name": "timerId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_satont_tsuwari_apps_api_internal_types.DOCApiValidationError"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_satont_tsuwari_apps_api_internal_types.DOCApiInternalError"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Timers"
                ],
                "summary": "Partially update timer",
                "parameters": [
                    {
                        "description": "Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_v1_timers.timerPatchDto"
                        }
                    },
                    {
                        "type": "string",
                        "description": "ID of channel",
                        "name": "channelId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of timer",
                        "name": "timerId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ChannelsTimers"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_satont_tsuwari_apps_api_internal_types.DOCApiValidationError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_satont_tsuwari_apps_api_internal_types.DOCApiInternalError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_satont_tsuwari_apps_api_internal_api_v1_settings_dashboardAccess.Entity": {
            "type": "object",
            "properties": {
                "channelId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "twitchUser": {
                    "$ref": "#/definitions/helix.User"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "github_com_satont_tsuwari_apps_api_internal_api_v1_settings_dashboardAccess.addUserDto": {
            "type": "object",
            "required": [
                "userName"
            ],
            "properties": {
                "userName": {
                    "type": "string"
                }
            }
        },
...
        "github_com_satont_tsuwari_apps_api_internal_types.DOCApiInternalError": {
            "type": "object",
            "properties": {
                "messages": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "github_com_satont_tsuwari_apps_api_internal_types.DOCApiValidationError": {
            "type": "object",
            "properties": {
                "messages": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "helix.User": {
            "type": "object",
            "properties": {
                "broadcaster_type": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "login": {
                    "type": "string"
                },
                "offline_image_url": {
                    "type": "string"
                },
                "profile_image_url": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "view_count": {
                    "type": "integer"
                }
            }
        },
...
        "internal_api_v1_timers.responseDto": {
            "type": "object",
            "required": [
                "isAnnounce",
                "text"
            ],
            "properties": {
                "isAnnounce": {
                    "type": "boolean"
                },
                "text": {
                    "type": "string",
                    "maxLength": 400
                }
            }
        },
        "internal_api_v1_timers.timerDto": {
            "type": "object",
            "required": [
                "enabled",
                "name",
                "responses"
            ],
            "properties": {
                "enabled": {
                    "type": "boolean"
                },
                "messageInterval": {
                    "type": "integer",
                    "maximum": 10000,
                    "minimum": 0
                },
                "name": {
                    "type": "string",
                    "maxLength": 510
                },
                "responses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_api_v1_timers.responseDto"
                    }
                },
                "timeInterval": {
                    "type": "integer",
                    "maximum": 120,
                    "minimum": 1
                }
            }
        },
        "internal_api_v1_timers.timerPatchDto": {
            "type": "object",
            "required": [
                "enabled"
            ],
            "properties": {
                "enabled": {
                    "type": "boolean"
                }
            }
        },
...
        "model.ChannelsTimers": {
            "type": "object",
            "properties": {
                "channelId": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "lastTriggerMessageNumber": {
                    "type": "integer"
                },
                "messageInterval": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "responses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ChannelsTimersResponses"
                    }
                },
                "timeInterval": {
                    "type": "integer"
                }
            }
        },
        "model.ChannelsTimersResponses": {
            "type": "object",
            "properties": {
                "isAnnounce": {
                    "type": "boolean"
                },
                "text": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "\"apiKey\" from /v1/profile response",
            "type": "apiKey",
            "name": "api-key",
            "in": "header"
        }
    }
}